sns-auto-poster/cloud-run/posting-engine/requirements.txt
flask==3.*
gunicorn==21.*
tweepy==4.*
requests==2.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
google-cloud-storage==2.*
google-cloud-aiplatform==1.*
vertexai==1.*
